chg: Make Generic
Make it so we can add new OIDCs as we need. Which requires
a rewrite.

BREAKING CHANGE: The refactoring of input variables is will break
current configurations that were using this module.
There has been no official release and this is less than 2 months
old; so it should not be to much of a disruption, if any.